Update Solution.cpp
I tried using a separate class for the graph, but then I realized that the AlphabetFinder class is basically a graph class, it has: the graph nodes, the graph, and the topological sorting function(s). It's just not a "general" graph class, and even if I had used a separate class for the graph, it wouldn't have been general either, because the way the graph is built is specific to the assignment's problem.
I've tried making the code a bit clearer.
If anyone wants to edit this, or use a separate graph for the class, they're welcome to. :)